Update eCommerceServiceDataAccessLayer as database schema has changed. Also switch to BigDecimal for longitude and latitude as its corresponding database colume has been changed from floating to decimal: https://stackoverflow.com/questions/159255/what-is-the-ideal-data-type-to-use-when-storing-latitude-longitude-in-a-mysql

+++ eCommerceService/src/main/java/com/bimelody/ecommerceservice/model/Store.java
@@ -3,6 +3,7 @@ package com.bimelody.ecommerceservice.model;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Builder
@@ -21,7 +22,7 @@ public class Store {
   private String instagramLink;
   private String twitterLink;
   private String storeLocation;
-  private Double storeLocationLatitude;
-  private Double storeLocationLongitude;
+  private BigDecimal storeLocationLatitude;
+  private BigDecimal storeLocationLongitude;
   private List<String> storeCategories;
 }
